final +head

+++ ft_atoi.c
@@ -1,16 +1,28 @@
+/* ************************************************************************** */
+/*                                                                            */
+/*                                                        :::      ::::::::   */
+/*   ft_atoi.c                                          :+:      :+:    :+:   */
+/*                                                    +:+ +:+         +:+     */
+/*   By: jgan <jgan@student.42.fr>                  +#+  +:+       +#+        */
+/*                                                +#+#+#+#+#+   +#+           */
+/*   Created: 2015/11/10 21:57:36 by jgan              #+#    #+#             */
+/*   Updated: 2015/11/10 22:05:52 by jgan             ###   ########.fr       */
+/*                                                                            */
+/* ************************************************************************** */
+
 #include "libft.h"
 
 int		ft_atoi(const char *str)
 {
-	int	tot;
-	int	sign;
-	int	i;
+	long	tot;
+	long	sign;
+	int		i;
 
 	tot = 0;
 	sign = 1;
 	i = 0;
 	while (str[i] == ' ' || str[i] == '\t' || str[i] == '\n' ||
-	str[i] == '\r' || str[i] == '\v' || str[i] == '\f')
+		str[i] == '\r' || str[i] == '\v' || str[i] == '\f')
 		i++;
 	if (str[i] == '+' || str[i] == '-')
 	{
@@ -23,5 +35,5 @@ int		ft_atoi(const char *str)
 		tot = tot * 10 + str[i] - '0';
 		i++;
 	}
-	return (tot * sign);
+	return ((int)(tot * sign));
 }
